ai-backend/pyproject.toml
[project]
name = "FastCrate"
version = "1.0.0"
description = "Base file structure of fast api"
license = "MIT"
authors = [
    {name = "mubashir-ahmed", email = "mubashirsidiki@gmail.com"}
]
readme = "README.md"

requires-python = ">=3.11,<3.12"
dependencies = [
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "gunicorn (>=23.0.0,<24.0.0) ; sys_platform == \"linux\"",
    "openai (>=1.78.1,<2.0.0)",
    "api-analytics[fastapi]>=1.2.7",
    "sqlalchemy[asyncio] (>=2.0.0,<3.0.0)",
    "asyncpg (>=0.29.0,<1.0.0)",
    "psycopg (>=3.0.0,<4.0.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)",
    "pgvector (>=0.2.0,<1.0.0)",
    "PyPDF2 (>=3.0.0,<4.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "numpy (>=1.24.0,<2.0.0)",
    "sentence-transformers>=2.0.0,<3.0.0",
    "redis>=4.5.0,<6.0.0",
    "aio-pika>=9.0.0,<10.0.0",
]
